backend/requirements.txt
fastapi
uvicorn[standard]
langchain
openai
faiss-cpu
sentence-transformers

tiktoken                  
transformers
google-generativeai
torch                     
python-dotenv             

httpx                     
tenacity                  

ipython
